refactor: 컨벤션 적용, 패키지 이동
